Arreglo del token e Insumos

+++ src/main/java/com/greencycle/ecoswap/ecoswap/repository/UsuarioRepository.java
@@ -1,10 +1,12 @@
 package com.greencycle.ecoswap.ecoswap.repository;
 
-import com.greencycle.ecoswap.ecoswap.model.Usuario;
-import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.greencycle.ecoswap.ecoswap.model.Usuario;
+
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     Optional<Usuario> findByEmail(String email);
     boolean existsByRuc(String ruc);
     boolean existsByEmail(String email);
